Lv 3-2) Calculator을 제네릭 클래스로. 하지만 객체에 연산자 적용이 안 돼서 double로 일일이 형변환을 했는데, 이게 맞는지 모르겠음..

+++ src/main/java/calculator/AddOperator.java
@@ -3,7 +3,7 @@ package calculator;
 public class AddOperator implements Operator{
     double result;
     @Override
-    public double operate(int num1, int num2){
+    public double operate(double num1, double num2){
         result = num1 + num2;
         return result;
     }
